pyproject.toml
[project]
name = "flare-rag-crawl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crawl4ai>=0.5.0.post4",
    "google-generativeai>=0.8.4",
    "pydantic-settings>=2.8.1",
    "structlog>=25.1.0",
]
